+++ src/Application/Local/Repository/LocalAppAuthRepositoryInterface.php
@@ -22,5 +22,5 @@ interface LocalAppAuthRepositoryInterface
 
     public function saveRenewedToken(RenewedAuthToken $renewedAuthToken): void;
 
-    public function save(LocalAppAuth $appAuth): void;
+    public function save(LocalAppAuth $localAppAuth): void;
 }